Merges OutOfBoundsMarker with Cursor
