+++ examples/multi-cursors/src/utils/__tests__/isOutOfBounds.test.ts
@@ -16,4 +16,14 @@ describe("isOutOfBounds", () => {
       isOutOfBounds({ x: 50, y: 50 }, { x: 0, y: 0, width: 100, height: 100 }),
     ).toBe(false);
   });
+
+  it("should return false if the position is inside the grace area", () => {
+    expect(
+      isOutOfBounds(
+        { x: 110, y: 110 },
+        { x: 0, y: 0, width: 100, height: 100 },
+        20,
+      ),
+    ).toBe(false);
+  });
 });
